chore: set up in memory saving of URL post requests

+++ internal/models/urlModels.go
@@ -10,5 +10,5 @@ type URLRecord struct {
 	URL           string `json:"url"`
 	Description   string `json:"description"`
 	CheckInterval int    `json:"checkInterval"`
-	Created_at    int    `json:"created_at"`
+	Created_at    int64  `json:"created_at"`
 }
